Try the output redirection thing AGAIN
Remove some xcopys. Note to self: don't overload the meaning of
GenerateProjectSpecificXyz
Investigate why the TestHost can't ProjectReference the
LocalTests_TermApp (it's becuase the xaml compiler tries to resolve a
winmd that just doesn't exist)
Investigate why there's a circular dependency from TerminalAzBridge but
only in parallel build.
Everything could be so nice if we could just finally land this