Switch Terminal to the WinAppSDK's "Hybrid CRT" model
"Hybrid CRT" is WinAppSDK's name for statically linking the C++ runtime
and dynamically linking the universal CRT (ucrt). It results in a minor
increase in binary sizes depending on how many STL classes each binary
uses, but overall it is cheaper than shipping a copy of the entire
runtime inside our package.

I have also updated to a new build of TerminalThemeHelpers that is built
in the same way.

Some of our dependencies still depend on the "app" versions of the
runtime libraries, so we are going to continue shipping the forwarders
in our package for their sake.

Because we're no longer shipping the desktop runtime, our package now
has a dependency on the VCLibs.UWPDesktop framework package. This is
not a problem on Windows 11 -- it looks like it's shipped inbox.

**BREAKING CHANGE**: When launched unpackaged, Terminal now requires the
vcruntime redist to be installed. If we can break XAML, Toolkit and
cpprest's dependency on the app CRT by switching them to a hybrid CRT
model we can become "self sufficient". It's not a priority right now.

Actually, back out the Hybrid CRT work in large part